flask_app/requirements.txt
flask==3.1.0
PyJWT==2.10.1
requests==2.32.3
pytest==8.3.5
pytest-html==4.1.1